Add `BaseHnswParam`. Fix sphinx warn.
